Refactor publisher search
I realize that I have a lot of repeating code, so I have to refactor in another pr
